feat: add functionality to serve files
