+++ Dockerfile
@@ -77,7 +77,8 @@ RUN ln -sf /usr/bin/clang /usr/bin/cc \
 RUN mkdir -p /opt/bin
 RUN ln -s /usr/bin/curl /opt/bin/curl
 
-COPY --from=0 /app/build/Chista_Asabru /bin/Chista_Asabru
+COPY --from=0 /app/build /bin
 COPY --from=0 /app/lib/asabru-handlers/build /asabru-handlers
 ENV PLUGINS_FOLDER_PATH=/asabru-handlers
+ENV PUBLIC_FOLDER_PATH=/bin/public
 CMD ["/bin/Chista_Asabru"]
