+++ automation_testing/ClickHouse/AsabruCH.py
@@ -26,6 +26,7 @@ class AsabruCH:
         self.table_name = config['table']
         self.port = config['port']
         self.ca_cert = config['ca_cert']
+        self.client_config = config['client_config']
         self.ssl = True if config['ssl']=='True' else False
 
         if self.ssl:
@@ -55,6 +56,7 @@ class AsabruCH:
             client.execute(f'CREATE DATABASE {self.database_name}')
             print('Database {self.database_name} created')
 
+        client.execute(f'DROP TABLE IF EXISTS {self.database_name}.cell_towers')
         client.disconnect()
 
         ## download data
@@ -80,9 +82,10 @@ class AsabruCH:
 
         ## Load data
         if not self.ssl:
-            os.system(f'clickhouse-client --host {self.ch_host} --user {self.ch_user} --password {self.ch_password} --port {self.port} --query "INSERT INTO {self.database_name}.cell_towers FORMAT CSVWithNames" < cell_towers.csv')
+            os.system(f'clickhouse client --host {self.ch_host} --user {self.ch_user} --password {self.ch_password} --port {self.port} --query "INSERT INTO {self.database_name}.cell_towers FORMAT CSVWithNames" < cell_towers.csv')
         else:
-            os.system(f'clickhouse-client --secure --host {self.ch_host} --user {self.ch_user} --password {self.ch_password} --port {self.port} --query "INSERT INTO {self.database_name}.cell_towers FORMAT CSVWithNames" < cell_towers.csv')
+            # os.system(f'clickhouse client --secure --host {self.ch_host} --user {self.ch_user} --password {self.ch_password} --port {self.port} --query "INSERT INTO {self.database_name}.cell_towers FORMAT CSVWithNames" < cell_towers.csv')
+            os.system(f'clickhouse client --secure --host {self.ch_host} --port {self.port} --config {self.client_config} --query "INSERT INTO {self.database_name}.cell_towers FORMAT CSVWithNames" < cell_towers.csv')
 
         print('Successfully loaded the data in to the table')
