+++ automation_testing/ClickHouse/ch_config.yaml
@@ -1,9 +1,10 @@
 host: 'localhost'
 username: 'default'
-password: '123456'
+password: ''
 port: '9120'
 ssl: 'True'
-ca_cert: '/home/vijay/SSL/custom_ca.crt'
+ca_cert: '/Users/midhundarvin/software/openssl/client.crt'
+client_config: '/Users/midhundarvin/software/openssl/clickhouse-client-ssl.xml'
 dataset_url: 'https://datasets.clickhouse.com/'
 dataset_file_name: 'cell_towers.csv.xz'
 dataset_csv_name: 'cell_towers.csv'
