--- config/config.xml
@@ -1,103 +1,0 @@
-<clickhouse-proxy-v2>
-
-	<protocol-server-config>
-		<protocol-server protocol="HTTP">
-			<protocol-port>8080</protocol-port>
-			<pipeline>ProtocolPipeline</pipeline>
-			<handler>CHttpProtocolHandler</handler>
-		</protocol-server>
-	</protocol-server-config>
-
-	<CLUSTERS>
-		<CLUSTER name="cluster1">
-			<END_POINTS>
-				<END_POINT name="ch_wire_http_node1">
-					<READ_WRITE>yes</READ_WRITE>
-					<PROXY_PORT>9100</PROXY_PORT>
-					<SERVICES>
-						<SERVICE name="ch_wire_service1">
-							<HOST>localhost</HOST>
-							<PORT>9000</PORT>
-							<PROTOCOL>CH_WIRE</PROTOCOL>
-							<PIPELINE>ClickHousePipeline</PIPELINE>
-							<HANDLER>CHWirePTHandler</HANDLER>
-						</SERVICE>
-						<SERVICE name="ch_http_service1">
-							<PROXY_PORT>9110</PROXY_PORT>
-							<PORT>8123</PORT>
-							<PROTOCOL>CH_HTTP</PROTOCOL>
-							<PIPELINE>ClickHousePipeline</PIPELINE>
-							<HANDLER>CHttpHandler</HANDLER>
-						</SERVICE>
-					</SERVICES>
-				</END_POINT>
-
-				<END_POINT name="ch_tls_node1">
-					<READ_WRITE>no</READ_WRITE>
-					<HOST>localhost</HOST>
-					<SERVICES>
-						<SERVICE name="ch_tls_wire_service1">
-							<PROXY_PORT>9120</PROXY_PORT>
-							<PORT>9440</PORT>
-							<PROTOCOL> CH_TLS_WIRE</PROTOCOL>
-							<PIPELINE>ClickHousePipeline</PIPELINE>
-							<HANDLER>CHWirePTHandler</HANDLER>
-						</SERVICE>
-						<SERVICE name="ch_tls_http_service1">
-							<PROXY_PORT>9130</PROXY_PORT>
-							<PORT>8443</PORT>
-							<PROTOCOL>CH_TLS_HTTP</PROTOCOL>
-							<PIPELINE>ClickHousePipeline</PIPELINE>
-							<HANDLER>CHttpsHandler</HANDLER>
-						</SERVICE>
-					</SERVICES>
-				</END_POINT>
-			</END_POINTS>
-		</CLUSTER>
-
-		<CLUSTER name="cluster2">
-			<END_POINTS>
-				<END_POINT name="pg_node1">
-					<READ_WRITE>no</READ_WRITE>
-					<HOST>localhost</HOST>
-					<SERVICES>
-						<SERVICE name="pg_wire_service1">
-							<PROXY_PORT>9140</PROXY_PORT>
-							<PORT>5432</PORT>
-							<PROTOCOL>PG_WIRE</PROTOCOL>
-							<PIPELINE>PostgreSQLPipeline</PIPELINE>
-							<HANDLER>CPostgreSQLHandler</HANDLER>
-						</SERVICE>
-						<SERVICE name="pg_tls_service1">
-							<PROXY_PORT>9150</PROXY_PORT>
-							<PORT>2345</PORT>
-							<PROTOCOL>PG_TLS_WIRE</PROTOCOL>
-							<PIPELINE>PostgreSQLPipeline</PIPELINE>
-							<HANDLER>CPostgreSQLHandler</HANDLER>
-						</SERVICE>
-					</SERVICES>
-				</END_POINT>
-				<END_POINT name="mysql_node1">
-					<READ_WRITE>yes</READ_WRITE>
-					<HOST>localhost</HOST>
-					<SERVICES>
-						<SERVICE name="mysql_wire_service1">
-							<PROXY_PORT>9160</PROXY_PORT>
-							<PORT>3306</PORT>
-							<PROTOCOL>MYSQL_WIRE</PROTOCOL>
-							<PIPELINE>MySQLPipeline</PIPELINE>
-							<HANDLER>CMySQLHandler</HANDLER>
-						</SERVICE>
-						<SERVICE name="mysql_tls_service1">
-							<PROXY_PORT>9170</PROXY_PORT>
-							<PORT>5501</PORT>
-							<PROTOCOL>MYSQL_TLS_WIRE</PROTOCOL>
-							<PIPELINE>MySQLPipeline</PIPELINE>
-							<HANDLER>CMySQLHandler</HANDLER>
-						</SERVICE>
-					</SERVICES>
-				</END_POINT>
-			</END_POINTS>
-		</CLUSTER>
-	</CLUSTERS>
-</clickhouse-proxy-v2>
