+++ src/config/intercepting-proxy-config.xml
