+++ src/config/serving-proxy-config.xml
