+++ src/pipelines/PassthroughPipeline.cpp
@@ -1,6 +1,5 @@
 #include "CProtocolSocket.h"
 #include "CProxySocket.h"
-#include "../test/ProxyInfo.h"
 #include "CClientSocket.h"
 #include "../config/ConfigSingleton.h"
 #include "ProtocolHelper.h"
