+++ src/pipelines/Pipeline.h
@@ -3,7 +3,6 @@
 #define PIPELINE_DOT_H
 #include "CProtocolSocket.h"
 #include "CProxySocket.h"
-#include "../test/ProxyInfo.h"
 #include "CClientSocket.h"
 #include "../config/ConfigSingleton.h"
 #include "ProtocolHelper.h"
