+++ src/pipelines/PostgreSQLPipeline.cpp
@@ -1,7 +1,6 @@
 
 #include "CProtocolSocket.h"
 #include "CProxySocket.h"
-#include "../test/ProxyInfo.h"
 #include "CClientSocket.h"
 #include "../config/ConfigSingleton.h"
 #include "ProtocolHelper.h"
