+++ src/public/index.html
@@ -1,0 +1,7 @@
+<!doctype html>
+<html>
+    <body>
+        <h1>Hello, world in an Html page</h1>
+        <p>A Paragraph</p>
+    </body>
+</html>
